Merged bigMessagesSupport into develop
NOTE: A change has been implemented in the "waitforBytes" functionality,
to make it simpler: Instead of calculating the Delay depending on
the bytes to read, we always wait for the same threshold. The original
ide has been proofed very tricky to test. The control on the Speed will
have to be carried out by a higher-layers of the App.