putting back in place the Events Queue in the EventQueueProcessor: Having a Queue allows us to avoid a blocking scenario that might happen if we use a Single Thread for the EventQueueProcessor. It also allows for running the Events in the EventBus in a non-blocking way, and we can use the Queue itself to develop some back-pressure in the future if needed.
The most important goal to achieve is to run the events published in the EventBus in a non-blocking way, but also without using a ot of Threads. using a Queue allows us to do that by using just 2 Threads: one for the Queue proccesing, and another for the Events consuming. If we want to increase performance in a particular handler, we can increase the number of threads in the EventsConsumer executor.